moves auxiliary functions to a new internal/astutils package
